--- database/models/organization.js
@@ -1,58 +1,0 @@
-'use strict';
-const {
-  Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class Organization extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The models/index file will call this method automatically.
-     */
-    static associate(models) {
-      // define association here
-    }
-  };
-  Organization.init({
-    id: {
-      allowNull: false,
-      autoIncrement: true,
-      primaryKey: true,
-      type: Sequelize.INTEGER
-    },
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    image: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    address: {
-      type: DataTypes.STRING,
-      allowNull: true
-    },
-    phone: {
-      type: DataTypes.INTEGER,
-      allowNull: true
-    },
-    email: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    welcomeText: {
-      type: DataTypes.TEXT,
-      allowNull: false
-    },
-    aboutText: {
-      type: DataTypes.TEXT,
-      allowNull: true
-    }
-  }, {
-    sequelize,
-    paranoid: true,
-    timestamps: true,
-    modelName: 'Organization',
-  });
-  return Organization;
-};
